third_party/WebKit/LayoutTests/inspector/sources/debugger-step/debugger-step-over-inlined-scripts.html

<html>
<head>

<script>
function f1()
{
    return 1; // Breakpoint.
}
f1();
</script>

<script>
function f2()
{
    return 2;
}
f2();
</script>

<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>

<script>

function test()
{
    var numberOfStepOver = 4;

    SourcesTestRunner.startDebuggerTest(step1, true);

    function step1()
    {
        SourcesTestRunner.showScriptSource("debugger-step-over-inlined-scripts.html", step2);
    }

    function step2(sourceFrame)
    {
        TestRunner.addResult("Script source was shown.");
        SourcesTestRunner.setBreakpoint(sourceFrame, 6, "", true);
        SourcesTestRunner.waitUntilPaused(step3);
        TestRunner.reloadPage(completeTest);
    }

    function step3()
    {
        var actions = ["Print"]; // First pause on breakpoint.
        for (var i = 0; i < numberOfStepOver; ++i)
            actions.push("StepOver", "Print");
        actions.push("Resume");
        SourcesTestRunner.waitUntilPausedAndPerformSteppingActions(actions);
    }

    function completeTest()
    {
        SourcesTestRunner.completeDebuggerTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that debugger StepOver will step through inlined scripts.
</p>
</body>
</html>
