third_party/WebKit/LayoutTests/inspector/sources/debugger-step/debugger-step-over-inlined-scripts-expected.txt

Tests that debugger StepOver will step through inlined scripts.

Script source was shown.
Call stack:
    0) f1 (debugger-step-over-inlined-scripts.html:7)
    1)  (debugger-step-over-inlined-scripts.html:9)

Executing StepOver...
Call stack:
    0) f1 (debugger-step-over-inlined-scripts.html:8)
    1)  (debugger-step-over-inlined-scripts.html:9)

Executing StepOver...
Call stack:
    0)  (debugger-step-over-inlined-scripts.html:1)

Executing StepOver...
Call stack:
    0)  (debugger-step-over-inlined-scripts.html:9)

Executing StepOver...
Call stack:
    0)  (debugger-step-over-inlined-scripts.html:12)

Executing StepOver...
Call stack:
    0)  (debugger-step-over-inlined-scripts.html:17)

Executing StepOver...
Call stack:
    0)  (debugger-step-over-inlined-scripts.html:1)

Executing StepOver...
Call stack:
    0)  (debugger-step-over-inlined-scripts.html:17)

Executing Resume...
Page reloaded.
